requirements.txt
django>=1.11.18
djangorestframework==3.9.0
django-rest-swagger==2.2.0
django-extensions==2.1.4
git+https://github.com/huifenqi/django-click.git
newrelic
elasticsearch-dsl>=6.0,<6.2
django-elasticsearch-dsl
